requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
redis
pydantic-settings
gunicorn
asyncpg
python-multipart
httpx
pgvector
numpy